services/embedding/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
transformers==4.37.0
pydantic==2.5.0
python-dotenv==1.0.0
torch>=2.0.0
httpx>=0.27.0
